Added HandleAck() to BasicMessage; implemented the new ack functionality into ParallelMessageHandler
